avoid to export PPolicyEnabledSlapdObject
