+++ Lib/slapdtest/__init__.py
@@ -7,9 +7,6 @@ See https://www.python-ldap.org/ for details.
 
 __version__ = '3.1.0'
 
-from slapdtest._slapdtest import (
-    SlapdObject, SlapdTestCase,
-    PPolicyEnabledSlapdObject,
-    SysLogHandler)
+from slapdtest._slapdtest import SlapdObject, SlapdTestCase, SysLogHandler
 from slapdtest._slapdtest import requires_ldapi, requires_sasl, requires_tls
 from slapdtest._slapdtest import skip_unless_ci
